# Go-2131 optimize badger memory consumption