I did all the benchmark in this way:
- Run desktop anytype app with 3700 objects in the objectstore (so it do the objectSearch under the hood and warm-up the cache)
- Run 1000 iterations of ObjectSearch with a filter that matches nothing

1. Main branch
  disk = SSD
  blockCache = 256mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=true
  compression = snappy
  **Result:
  4518ms for 1000 iterations**

2. Main branch + disabled valuePrefetch for iterator
  disk = SSD
  blockCache = 256mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=false
  compression = snappy
  **Result:
  1948ms for 1000 iterations**

3. all badger options tunes(this branch)
  disk = SSD
  blockCache = 0mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=false
  compression = none
  **Result:
  1761ms for 1000 iterations**

4. This branch on HDD
  disk = HDD (to check that disabled cache doesn't affect the speed)
  cache = 0mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=false
  compression = none
  **Result:
  1254ms for 1000 iterations (probably db files were fully cached by OS)**

Allocs test:
- Start desktop app (hot or cold account start)
1. This branch
  disk = SSD
  blockCache = 0mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=false
  compression = none
  **Result:
  Cold start(fetch from remote): 959MB
  Hot start: 313MB**
 
2. Main branch
  disk = SSD
  blockCache = 256mb
  valuethreshold = 1024*1024
  iterator.valuePrefetch=true
  compression = snappy
  **Result:
  Cold start(fetch from remote): 1906MB
  Hot start: 352MB**